# docs/openapi.yaml
openapi: 3.0.1
info:
  title: MTGLibrary API
  version: '1.0'
  description: API for MTGLibrary — endpoints for cards, collection management, decks, and users.
servers:
  - url: http://localhost:3000
paths:
...
  /collection:
    get:
      summary: List User Collection
      parameters:
        - in: query
          name: deck_id
          schema:
            type: string
          description: Filter by deck ('null' or 'binder' for unassigned)
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: type_line
          schema:
            type: string
      responses:
        '200':
          description: List of user cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCard'
    post:
      summary: Add Single Card to Collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCardInput'
      responses:
        '201':
          description: Created
...
  /decks:
    get:
      summary: List Decks
      responses:
        '200':
          description: User decks
    post:
      summary: Create Deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                commander:
                  type: object
      responses:
        '201':
          description: Created
  /decks/import:
    post:
      summary: Smart Deck Import
      description: Import a deck and optionally add missing cards to collection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deck:
                  type: object
                  properties:
                    name:
                      type: string
                    commander:
                      type: object
                cards:
                  type: array
                  items:
                    type: object
                options:
                  type: object
                  properties:
                    checkCollection:
                      type: boolean
                    addToCollection:
                      type: boolean
      responses:
        '200':
          description: Import success details
  /decks/{id}:
    get:
      summary: Get Deck Details & Cards
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deck object and items list
    put:
      summary: Update Deck Metadata
    delete:
      summary: Delete Deck
  /decks/{id}/cards/batch:
    post:
      summary: Batch Add Cards to Deck
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cards:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserCardInput'
      responses:
        '200':
          description: Added

  # --- USERS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        count:
          type: integer
        deck_id:
          type: integer
          nullable: true
        is_wishlist:
          type: boolean
        data:
          type: object
    UserCardInput:
      type: object
      properties:
        scryfall_id:
          type: string
        name:
          type: string
        count:
          type: integer
        finish:
          type: string
        deck_id:
          type: integer
        is_wishlist:
          type: boolean
        data:
          type: object